:sparkles: feat: delete useless codes
