+++ src/containers/Home/Announcement/Announcement.tsx
@@ -9,7 +9,7 @@ import TableWrapper from 'components/TableWrapper/TableWrapper'
 import Poppers from 'components/Poppers/Poppers'
 import Loading from 'components/Loading/Loading'
 import EditModal from './components/EditModal/EditModal'
-import ConfirmModal from './components/ConfirmModal/ConfirmModal'
+import ConfirmModal from '../../../components/ConfirmModal/ConfirmModal'
 import { formatISODate } from 'shared/utils'
 import { Props } from './Announcement.connect'
 import styles from './Announcement.module.scss'
